requirements.txt
cython
pandas
numpy
matplotlib
adafruit_circuitpython_bno055
adafruit-circuitpython-motorkit
adafruit-circuitpython-motor
scipy
skyfield
pytz
spktype01
jplephem
networkx
spiceypy
ephem
geomag
luma.core
luma.oled
luma.emulator
pillow
tzwhere
